Rename Attribute concept to Record throughout codebase
Renames src/attributes/ → src/records/, all Attribute classes to Record
equivalents (KeyValueRecord, LongTextRecord, BlobRecord, RecordsRepository),
and adds /records/* HTTP endpoints alongside the existing /attributes/* routes.

Backwards compatibility is preserved: client.Attribute getter, old class name
aliases (KeyValueAttribute etc.), old HTTP routes, and wire protocol strings
('KeyValueAttribute' in $hasDataType queries) all remain unchanged.
