+++ src/records/abstract/is_serializable.ts
