+++ src/records/abstract/serialized_change_with_metadata.ts
