+++ src/records/errors/authorization_error.ts
